[bugfix] 광물 제출 시 Progressbar 적용 안되는 문제 해결

+++ Source/AbyssDiverUnderWorld/Interactable/OtherActors/ADDroneSeller.h
@@ -133,7 +133,7 @@ private:
 		}
 
 		CurrentMoney = NewCurrentMoney;
-		MoneyRatio = CurrentMoney / TargetMoney;
+		MoneyRatio = (float)CurrentMoney / TargetMoney;
 		OnRep_CurrentMoney();
 	}
